Add wave model tutorial + data sources + move GEM utils + restore old masking method
