+++ nsbtools/data/included_data.csv
@@ -1,14 +1,14 @@
-DATA_TYPE,SPECIES,TEMPLATE,DENSITY,HEMI,DESCRIPTION
-surf,human,fsLR,4k,L,Cortical midthickness surface mesh
-surf,human,fsLR,4k,R,Cortical midthickness surface mesh
-surf,human,fsLR,32k,L,Cortical midthickness surface mesh
-surf,human,fsLR,32k,R,Cortical midthickness surface mesh
-surf,macaque,fsLR,10k,L,Cortical midthickness surface mesh
-surf,marmoset,fsLR,10k,L,Cortical midthickness surface mesh
-medmask,human,fsLR,4k,L,Medial wall mask for midthickness
-medmask,human,fsLR,4k,R,Medial wall mask for midthickness
-medmask,human,fsLR,32k,L,Medial wall mask for midthickness
-medmask,human,fsLR,32k,R,Medial wall mask for midthickness
-medmask,macaque,fsLR,10k,L,Medial wall mask for midthickness
-medmask,marmoset,fsLR,10k,L,Medial wall mask for midthickness
-fcgradient1,human,fsLR,32k,L,First functional connectivity gradient from Margulies et al. 2016
+DATA_TYPE,SPECIES,TEMPLATE,DENSITY,HEMI,SOURCE,DESCRIPTION
+surf,human,fsLR,4k,L,Glasser et al. (2011) J Neurosci,Cortical midthickness surface mesh (Conte69)
+surf,human,fsLR,4k,R,Glasser et al. (2011) J Neurosci,Cortical midthickness surface mesh (Conte69)
+surf,human,fsLR,32k,L,Glasser et al. (2011) J Neurosci,Cortical midthickness surface mesh (Conte69)
+surf,human,fsLR,32k,R,Glasser et al. (2011) J Neurosci,Cortical midthickness surface mesh (Conte69)
+surf,macaque,fsLR,10k,L,Donahue et al. (2016) J Neurosci,Cortical midthickness surface mesh (Yerkes19)
+surf,marmoset,fsLR,10k,L,Liu et al. (2021) NeuroImage,Cortical midthickness surface mesh (MBMv3)
+medmask,human,fsLR,4k,L,Glasser et al. (2016) J Neurosci,Medial wall mask for midthickness
+medmask,human,fsLR,4k,R,Glasser et al. (2016) J Neurosci,Medial wall mask for midthickness
+medmask,human,fsLR,32k,L,Glasser et al. (2016) J Neurosci,Medial wall mask for midthickness
+medmask,human,fsLR,32k,R,Glasser et al. (2016) J Neurosci,Medial wall mask for midthickness
+medmask,macaque,fsLR,10k,L,Donahue et al. (2016) J Neurosci,Medial wall mask for midthickness
+medmask,marmoset,fsLR,10k,L,Liu et al. (2021) NeuroImage,Medial wall mask for midthickness
+fcgradient1,human,fsLR,32k,L,Margulies et al. (2016) PNAS,First functional connectivity gradient
