+++ nsbtools/data/sp-human_tpl-fsLR_den-32k_hemi-L_midthickness.surf.gii
