+++ nsbtools/data/sp-human_tpl-fsLR_den-32k_hemi-R_medmask.label.gii
