+++ nsbtools/data/sp-human_tpl-fsLR_den-4k_hemi-L_midthickness.surf.gii
