+++ nsbtools/data/sp-human_tpl-fsLR_den-4k_hemi-R_midthickness.surf.gii
