+++ nsbtools/data/sp-marmoset_tpl-fsLR_den-10k_hemi-L_midthickness.surf.gii
